requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
filelock==3.12.2
identify==2.5.26
idna==3.4
nodeenv==1.8.0
oauthlib==3.2.2
platformdirs==3.10.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.3
virtualenv==20.24.2
